Refactor round fulfillment and response.
